+++ packages/react/src/icons/app/VolumeHigh.tsx
@@ -1,0 +1,38 @@
+import type { SVGProps } from "react"
+import { Ref, forwardRef } from "react"
+const SvgVolumeHigh = (
+  props: SVGProps<SVGSVGElement>,
+  ref: Ref<SVGSVGElement>
+) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    viewBox="0 0 24 24"
+    ref={ref}
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      d="M11 5L6 9H3V15H6L11 19V5Z"
+      vectorEffect="non-scaling-stroke"
+    />
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      d="M15 9.5C16 10.6667 16 13.3333 15 14.5"
+      vectorEffect="non-scaling-stroke"
+    />
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      d="M17.5 7C19.6667 9.33333 19.6667 14.6667 17.5 17"
+      vectorEffect="non-scaling-stroke"
+    />
+  </svg>
+)
+const ForwardRef = forwardRef(SvgVolumeHigh)
+export default ForwardRef